# swagger.yaml
openapi: 3.0.3
info:
  title: Smart Campus - Core Business API (B6)
  description: API phục vụ quản lý hồ sơ học viên, điểm danh và tra cứu điểm số.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/attendances/bulk:
    post:
      summary: Diem danh hang loat cho mot lop trong mot ngay
      tags: [Business - Attendances]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                classId:
                  type: integer
                  example: 10
                attendanceDate:
                  type: string
                  format: date
                  example: "2026-06-10"
                students:
                  type: array
                  items:
                    type: object
                    properties:
                      studentId:
                        type: integer
                        example: 1
                      status:
                        type: string
                        example: "Present"
                      note:
                        type: string
      responses:
        '201':
          description: Diem danh hang loat thanh cong
        '400':
          description: Du lieu gui len khong hop le
        '404':
          description: Khong tim thay hoc vien
        '409':
          description: Du lieu bi trung hoac hoc vien chua dang ky lop
  /api/v1/attendances/summary/students/{studentId}/classes/{classId}:
    get:
      summary: Tinh ty le phan tram chuyen can cua hoc vien
      tags: [Business - Attendances]
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Summary chuyen can
        '404':
          description: Khong tim thay hoc vien
        '409':
          description: Hoc vien chua dang ky lop
...